8/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.8
django-braces==1.15.0
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
django-social-auth==0.7.28
django-sslserver==0.22
djangorestframework==3.12.4
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
oauth2==1.9.0.post1
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-openid==2.2.5
python-social-auth==0.2.21
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
